% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.sparse.mediation.R
\name{cv.sparse.mediation}
\alias{cv.sparse.mediation}
\title{Conduct K-fold cross validation for sparse medication with elastic net with multiple tuning parameters}
\usage{
cv.sparse.mediation(X, M, Y, tol = 10^(-10), K = 5, max.iter = 100,
  lambda = log(1 + (1:15)/50), alpha = (0:4)/4, figure = NULL,
  multicore = 1, seednum = 1e+06, display = TRUE)
}
\arguments{
\item{X}{One-dimensional predictor}

\item{M}{Multivariate mediator}

\item{Y}{Outcome}

\item{tol}{(default -10^(-10)) convergence criterion}

\item{K}{(default=5) number of cross-validation folds}

\item{max.iter}{(default=100) maximum iteration}

\item{lambda}{(default=log(1+(1:30)/100)) tuning parameter for L1 penalization}

\item{alpha}{(defult=c(1:4)/4) tuning parameter for L2 penalization}

\item{figure}{(defult=NULL) print figures for mean predictive errors by tuning parameters alpha and lambda}

\item{multicore}{(default=1) number of multicore}

\item{seednum}{(default=10000) seed number for cross validation}

\item{glmnet.penalty.factor}{(default=c(0,rep(1,2*V))) give different weight of penalization for the 2V mediation paths.}

\item{disply}{devault=FALSE}
}
\value{
re list of sparse.mediation per each alpha
}
\description{
Conduct K-fold cross validation for sparse medication with elastic net with multiple tuning parameters
}
\examples{
N=100
V=50
set.seed(1234)
a = rbinom(V,1,0.1)*5;b<-a
X = rnorm(N)
M =  X \%*\% t(a)+ matrix(rnorm(N*V),N,V)
Y =  X + M \%*\% b + rnorm(N)
cv.sparse.mediation(X,M,Y,tol=10^(-10),max.iter=100,lambda = log(1+(1:15)/50))
}
\references{
TBA
}
\author{
Seonjoo Lee, \email{sl3670@cumc.columbia.edu}
}
\keyword{glmnet}
\keyword{hdlfpca}
