% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.mediation.largep.R
\name{sparse.mediation.largep}
\alias{sparse.mediation.largep}
\title{Conduct sparse mediation for parge p ( p > n) with L1 or L2 penalization using fast computation of inverse matrix}
\usage{
sparse.mediation.largep(X, M, Y, tol = 10^(-10), max.iter = 10,
  lambda = log(1 + (1:50)/125), glmnet.penalty.factor = c(0, rep(1, 2 * V)),
  alpha = 1, threshold = 1e-05, verbose = FALSE)
}
\arguments{
\item{X}{One-dimensional predictor}

\item{M}{Multivariate mediator}

\item{Y}{Outcome}

\item{tol}{(default -10^(-10)) convergence criterion}

\item{max.iter}{(default=100) maximum iteration}

\item{lambda}{(default=log(1+(1:50)/125)) tuning parameter for L1 penalization}

\item{glmnet.penalty.factor}{(default=c(0,rep(1,2*V))) give different weight of penalization for the 2V mediation paths.}

\item{alpha}{(defult=1) tuning parameter for L2 penalization}

\item{figure}{(defult=NULL) print figures for mean predictive errors by tuning parameters alpha and lambda}
}
\value{
c directeffect

hatb Path b (M->Y given X) estimates

hata Path a (X->M) estimates

medest Mediation estimates (a*b)

alpha

lambda

nump Number of selected mediation paths
}
\description{
Fit a mediation model via penalized maximum likelihood and structural equation model.
The regularization path is computed for the lasso or elasticnet penalty at a grid of
values for the regularization parameter lambda. Currently, mediation analysis is developed based on gaussian assumption.
}
\details{
Multiple Mediaton Model:
(1) M = Xa + e1
(2) Y = Xc' + Mb + e2
And in the optimization, we do not regularize c', due to the assumption of partial mediation.
}
\examples{
N=100
V=500
set.seed(1234)
    covmat=matrix(0,V+2,V+2);
    covmat[1,2]=0.5;covmat[1, (1:3)+2]=rep(r.a,3);covmat[2, (1:3)+2]=rep(r.b,3);
    covmat=covmat+t(covmat);diag(covmat)<-1
    sqrtmat = sqrtmat.comp(covmat)
    tmpmat = matrix(rnorm(N*(V+2)),N,V+2) \%*\% sqrtmat
    X=tmpmat[,1]
    Y=tmpmat[,2]
    M=tmpmat[,-c(1:2)]
    sparse.mediation.largep(X,M,Y,tol=10^(-10),max.iter=100,lambda = log(1+(1:25)/50))

}
\references{
TBA
}
\author{
Seonjoo Lee, \email{sl3670@cumc.columbia.edu}
}
\keyword{L1penalization}
\keyword{highdimensional}
\keyword{mediation}
