% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.mediation.twoway.R
\name{sparse.mediation.twoway}
\alias{sparse.mediation.twoway}
\title{Conduct sparse mediation with elastic net with multiple tuning parameters}
\usage{
sparse.mediation.twoway(X, M, Y, tol = 10^(-10), max.iter = 100,
  lambda = log(1 + (1:30)/100), alpha = (1:4)/4,
  glmnet.penalty.factor = c(0, rep(1, 2 * V)))
}
\arguments{
\item{X}{One-dimensional predictor}

\item{M}{Multivariate mediator}

\item{Y}{Outcome}

\item{tol}{(default -10^(-10)) convergence criterion}

\item{max.iter}{(default=100) maximum iteration}

\item{lambda}{(default=log(1+(1:30)/100)) tuning parameter for L1 penalization}

\item{alpha}{(defult=c(1:4)/4) tuning parameter for L2 penalization}

\item{glmnet.penalty.factor}{(default=c(0,rep(1,2*V))) give different weight of penalization for the 2V mediation paths.}

\item{figure}{(defult=NULL) print figures for mean predictive errors by tuning parameters alpha and lambda}
}
\value{
re list of sparse.mediation per each alpha
}
\description{
Conduct sparse mediation with elastic net with multiple tuning parameters
}
\examples{
N=100
V=50
set.seed(1234)
a = rbinom(V,1,0.1)*5;b<-a
X = rnorm(N)
M =  X \%*\% t(a)+ matrix(rnorm(N*V),N,V)
Y =  X + M \%*\% b + rnorm(N)
sparse.mediation.twoway(X,M,Y,tol=10^(-10),max.iter=100,lambda = log(1+(1:25)/50))
}
\references{
TBA
}
\author{
Seonjoo Lee, \email{sl3670@cumc.columbia.edu}
}
\keyword{glmnet}
\keyword{hdlfpca}
