% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrtmat.comp.R
\name{sqrtmat.comp}
\alias{sqrtmat.comp}
\title{Compute squareroot matrix}
\usage{
sqrtmat.comp(mat, thresh = 10^(-20))
}
\arguments{
\item{mat}{One-dimensional predictor}
}
\value{
squareroot of matrix
}
\description{
Compute squareroot matrix
}
\examples{
set.seed(1234)
x=matrix(100*20,100,20)
covmat=cov(x)
sqrtmat.comp(covmat)

}
\references{
TBA
}
\author{
Seonjoo Lee, \email{sl3670@cumc.columbia.edu}
}
\keyword{L1penalization}
\keyword{highdimensional}
\keyword{mediation}
