% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.sparse.mediation.R
\name{cv.sparse.mediation}
\alias{cv.sparse.mediation}
\title{Conduct K-fold cross validation for sparse mediation with elastic net with multiple tuning parameters}
\usage{
cv.sparse.mediation(X, M, Y, tol = 10^(-10), K = 5, max.iter = 100,
  lambda = log(1 + (1:15)/50), lambda2 = c(0.2, 0.5), alpha = 1,
  tau = c(0.5, 1, 2), multicore = 1, seednum = 1e+06, verbose = FALSE)
}
\arguments{
\item{X}{One-dimensional predictor}

\item{M}{Multivariate mediator}

\item{Y}{Outcome}

\item{tol}{(default -10^(-10)) convergence criterion}

\item{K}{(default=5) number of cross-validation folds}

\item{max.iter}{(default=100) maximum iteration}

\item{lambda}{(default=log(1+(1:30)/100)) tuning parameter for L1 penalization}

\item{lambda2}{(default=c(0.2,0.5)) tuning parameter for inverse covariance matrix sparsity. Only used if n>(2*V).}

\item{alpha}{(defult=1) tuning parameter for L2 penalization}

\item{tau}{(default=1) tuning parameter for differential weight for L1 penalty.}

\item{multicore}{(default=1) number of multicore}

\item{seednum}{(default=10000) seed number for cross validation}

\item{verbose}{(default=FALSE)}
}
\value{
cv.lambda: optimal lambda

cv.tau: optimal tau

cv.alpha: optimal tau

cv.mse: minimum MSE value

mse: Array of MSE, length(alpha) x length(lambda) x length (tau)

lambda: vector of lambda

tau: vector of tau used

alpha: vector of alpha used

z: cross-valication results
}
\description{
Conduct K-fold cross validation for sparse mediation with elastic net with multiple tuning parameters
}
\examples{
library(sparsemediation)
N=200
V=50
set.seed(1234)
a = rep(0,V);a[1:3]<-1;b<-a
X = rnorm(N)
M =  X \%*\% t(a)+ matrix(rnorm(N*V),N,V)
Y =  as.vector(X + M \%*\% b + rnorm(N))
cvfit<-cv.sparse.mediation(X, M, Y, K = 4, lambda=c(0.05,0.2,length=5), tau=c(0.5,1,2),multicore = 4)
cvfit$cv.lambda
cvfit$cv.tau
fit<-sparse.mediation(X,M,Y,lambda = cvfit$cv.lambda, alpha=cvfit$cv.alpha,tau=cvfit$cv.tau)
}
\references{
TBA
}
\author{
Seonjoo Lee, \email{sl3670@cumc.columbia.edu}
}
\keyword{glmnet}
\keyword{hdlfpca}
