% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.mediation.R
\name{sparse.mediation}
\alias{sparse.mediation}
\title{Conduct sparse mediation with elastic net}
\usage{
sparse.mediation(X, M, Y, tol = 10^(-5), max.iter = 50, lambda = log(1 +
  (1:30)/100), lambda2 = c(0.3, 0.5), alpha = 1, tau = 1,
  verbose = FALSE, Omega.out = FALSE)
}
\arguments{
\item{X}{One-dimensional predictor}

\item{M}{Multivariate mediator}

\item{Y}{Outcome}

\item{tol}{(default -10^(-5)) convergence criterion}

\item{max.iter}{(default=100) maximum iteration}

\item{lambda}{(default=log(1+(1:50)/125)) tuning parameter for L1 penalization}

\item{lambda2}{(default=c(0.2,0.5)) tuning parameter for L1 penalization for covariance matrix, used only for p>n.}

\item{alpha}{(default=1) tuning parameter for L2 penalization}

\item{tau}{(default=1) tuning parameter for differentail weight between paths a (X -> M) and b (M -> Y)}

\item{verbose}{(default=TRUE) print progress.}

\item{Omega.out}{(defult=TRUE) output Omega estimates}
}
\value{
c: directeffect per each tuning parameter lambda. length(lambda)-dimensional vector

hatb: Path b (M->Y given X) estimates: V-by-lenbth(lambda)  matrix

hata: Path a (X->M) estimates: V-by-lenbth(lambda)  matrix

medest: Mediation estimates (a*b): V-by-lenbth(lambda)  matrix

alpha: a scalor of the numing parameter for L2 regularization

lambda: a vector of tuning parameters for L1-penalization

tau: weight used.

nump: Number of selected mediation paths

Omega Estimated covariance matrix of the mediator
}
\description{
Fit a mediation model via penalized maximum likelihood and structural equation model.
The regularization path is computed for the lasso or elasticnet penalty at a grid of
values for the regularization parameter lambda. Currently, mediation analysis is developed based on gaussian assumption.
}
\details{
Multiple Mediaton Model:
(1) M = Xa + e1
(2) Y = Xc' + Mb + e2
And in the optimization, we do not regularize c', due to the assumption of partial mediation.
}
\examples{
library(sparsemediation)
N=100
V=50
set.seed(1234)
a = rep(0,V);a[1:3]<-5;b<-a
X = rnorm(N)
M =  X \%*\% t(a)+ matrix(rnorm(N*V),N,V)
Y =  X + M \%*\% b + rnorm(N)
sparse.mediation(X,M,Y)
}
\references{
TBA
}
\author{
Seonjoo Lee, \email{sl3670@cumc.columbia.edu}
}
\keyword{glmnet}
\keyword{highdimensional}
\keyword{mediation}
