% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrtmat.comp.R
\name{sqrtmat.comp}
\alias{sqrtmat.comp}
\title{Compute squareroot of large covariance matrix}
\usage{
sqrtmat.comp(mat, thresh = 10^(-20), K = NULL)
}
\arguments{
\item{mat}{Covariance matrix}

\item{K}{Dimension of the matrix. Default is set as ncol(mat)-1}
}
\value{
squareroot of matrix
}
\description{
Compute squareroot of large covariance matrix
}
\examples{
set.seed(1234)
x=matrix(100*20,100,20)
covmat=cov(x)
sqrtmat.comp(covmat)

}
\references{
TBA
}
\author{
Seonjoo Lee, \email{sl3670@cumc.columbia.edu}
}
\keyword{squareroot}
